Merge pull request #81 from rster2002/milestone

+++ assets/components/dnd/js/characterEditor.js
@@ -1,5 +1,13 @@
 sUid = sessionStorage.getItem("::uid");
 
+timer = setInterval(function() {
+	console.log("timer fire");
+	if (sessionStorage.getItem("::saved") !== "false") {
+		saveCharacter();
+	}
+}, 15000);
+
+
 $(".innerPage").ready(() => {
 	$(".characterContainer").load("../assets/components/dnd/pages/characterSheet.html");
 });
@@ -12,6 +20,8 @@ sessionStorage.setItem("::saved","false");
 console.log("hmm");
 function saveCharacter() {
     
+	se = false;
+	
 	try {
 		s();
 		console.log(characterObj);
@@ -22,9 +32,21 @@ function saveCharacter() {
 		} else {
 			var i = prompt("Type a name for this caracter sheet");
 			if (i) {
-				sessionStorage.setItem("::saved",i);
-				dbUsers.child(sUid).child("characters").child(i).set(characterObj);
-				note.open("Saved " + sessionStorage.getItem("::saved"), 1000);
+				if (sessionStorage.getItem("::saved") === "false") {
+					dbUsers.child(sUid).child("characters").once("value", (e) => {
+						if (e.hasChild(i)) {
+							se = confirm("Are you sure you want to overwrite this character sheet?");
+						}
+					});
+				} else {
+					se = true;
+				}
+				
+				if (se) {
+					sessionStorage.setItem("::saved",i);
+					dbUsers.child(sUid).child("characters").child(i).set(characterObj);
+					note.open("Saved " + sessionStorage.getItem("::saved"), 1000);
+				}
 			}
 		}
 	} catch(e) {
@@ -33,13 +55,28 @@ function saveCharacter() {
 }
 
 function saveAsCharacter() {
+	
+	se = false;
+	
 	s();
 	var i = prompt("Type a name for this caracter sheet");
 	if (i) {
-		loader.show();
-		sessionStorage.setItem("::saved",i);
-		dbUsers.child(sUid).child("characters").child(i).set(characterObj);
-		loader.hide();
+		if (sessionStorage.getItem("::saved") === "false") {
+			dbUsers.child(sUid).child("characters").once("value", (e) => {
+				if (e.hasChild(i)) {
+					se = confirm("Are you sure you want to overwrite this character sheet?");
+				}
+			});
+		} else {
+			se = true;
+		}
+		
+		if (se) {
+			loader.show();
+			sessionStorage.setItem("::saved",i);
+			dbUsers.child(sUid).child("characters").child(i).set(characterObj);
+			loader.hide();
+		}
 	}
 }
